azure work

+++ app/src/main/java/com/zygr/easyprotein/app/ModifyFoodEntry.java
@@ -2,7 +2,7 @@ package com.zygr.easyprotein.app;
 
 
 public interface ModifyFoodEntry {
-    public void addFoodEntry(double calorie, double protein);
+    public void addFoodEntry(int calorie, int protein);
     public void resetFoodEntry();
     public void confirmReset();
     public void dataModified();
